LayoutTests/fast/text-autosizing/ios/idempotentmode/viewport-change-relayout-expected.txt

PASS em.getBoundingClientRect().x + em.getBoundingClientRect().width - 0.0001 <= tail.getBoundingClientRect().x is true
PASS successfullyParsed is true

TEST COMPLETE
Japan with Everything about modern and traditional emphasis
